Create ex1incomecal.cpp
Write a C++ program to calculate the Net-Pay of an employee as follow after deduction
of pension (7%) and tax. The program should read the gross salary, worked hours, and
income tax rate. If the employee worked hours exceed 40 hrs., the program prompts
the user to enter an over-time bonus rate/hour and find the bonus payment by
multiplying the extra hours worked with the provided bonus rate.
 net Salary = (gross Salary – pension – income tax) + Overtime payment